Last Update

+++ SkillSale/Areas/Identity/Pages/Account/Manage/ProfileImage.cshtml.cs
@@ -6,7 +6,10 @@ using SkillSale.Areas.Identity.Data;
 using System.Drawing.Imaging;
 using System.Drawing;
 using SkillSale.Data;
-
+using Microsoft.EntityFrameworkCore;
+using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.Processing;
+using System.IO;
 
 namespace SkillSale.Areas.Identity.Pages.Account.Manage
 {
@@ -15,9 +18,13 @@ namespace SkillSale.Areas.Identity.Pages.Account.Manage
         private readonly UserManager<SkillSaleUser> _userManager;
         public string profileImageUrl;
 
+        [TempData]
+        public string StatusMessage { get; set; }
+
+
         private readonly SkillSaleContext _context;
 
-        public ProfileImageModel(SkillSaleContext context,  UserManager<SkillSaleUser> userManager){
+        public ProfileImageModel(UserManager<SkillSaleUser> userManager, SkillSaleContext context  ){
             _userManager = userManager;  
             _context = context;
         }
@@ -30,10 +37,24 @@ namespace SkillSale.Areas.Identity.Pages.Account.Manage
             return Page();
         }
 
-        public async Task OnPost(IFormFile file) {
-            if (file != null && file.Length > 0) {
+        public async Task OnPost(IFormFile file)
+        {
+            if (file != null && file.Length > 0)
+            {
                 try
                 {
+                    var allowedExtensions = new[] { ".jpg", ".jpeg", ".png", ".gif" };
+                    var ext = Path.GetExtension(file.FileName).ToLowerInvariant();
+
+                    if (!allowedExtensions.Contains(ext))
+                    {
+                        profileImageUrl = "profilePictures/noimage.jpg";
+                        StatusMessage = ("Файл не является изображением.");
+                        RedirectToPage();
+                        throw new Exception("Загруженный файл в картинку профиля, не является картинкой");
+                        
+                    }
+
                     var currentUser = await _userManager.GetUserAsync(User);
 
                     var deleteFilename = currentUser.ProfileImage;
@@ -47,11 +68,25 @@ namespace SkillSale.Areas.Identity.Pages.Account.Manage
                     await Console.Out.WriteLineAsync(delPath);
                     Directory.CreateDirectory(Path.GetDirectoryName(savePath));
 
-                    if (System.IO.File.Exists(delPath))
-                        System.IO.File.Delete(delPath);
+                    string[] staticImageFiles = ["female.webp", "male.png", "noimage.jpg"];
+                    
+                    if (System.IO.File.Exists(delPath)) 
+                        if (staticImageFiles.Contains(deleteFilename))
+                            System.IO.File.Delete(delPath);
 
-                    using (var stream = new FileStream(savePath, FileMode.Create)) {
-                        await file.CopyToAsync(stream);
+                    using (var stream = new FileStream(savePath, FileMode.Create))
+                    {
+                        // Изменение размера изображения до заданных размеров
+                        using (var image = SixLabors.ImageSharp.Image.Load(file.OpenReadStream()))
+                        {
+                            var options = new ResizeOptions
+                            {
+                                Mode = ResizeMode.Max, // Изменение размера с сохранением пропорций
+                                Size = new SixLabors.ImageSharp.Size(300, 300) // Заданные размеры (ширина, высота)
+                            };
+                            image.Mutate(x => x.Resize(options));
+                            image.SaveAsJpeg(stream); // Сохранение изображения в формате JPEG
+                        }
                     }
 
                     profileImageUrl = "profilePictures/" + filename;
@@ -61,11 +96,9 @@ namespace SkillSale.Areas.Identity.Pages.Account.Manage
                 }
                 catch (Exception ex)
                 {
-                    await Console.Out.WriteLineAsync( ex.Message);
+                    await Console.Out.WriteLineAsync(ex.Message);
                 }
             }
         }
-
-       
     }
 }
